# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "Tạo tài khoản mới, chưa đăng nhập tự động.",
        "operationId": "authRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đăng ký thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Đăng ký thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicAuthUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu validation không hợp lệ hoặc thiếu Turnstile token khi Turnstile bật. Turnstile missing token trả code TURNSTILE_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "CSRF failure trả code CSRF_INVALID. Turnstile verification/config/provider failure trả code TURNSTILE_FAILED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email đã được sử dụng.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Vượt quá giới hạn đăng ký trong cửa sổ throttle.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Đăng ký tài khoản",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Tạo phiên đăng nhập và lưu token vào cookie.",
        "operationId": "authLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Đăng nhập thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicAuthUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu validation không hợp lệ hoặc thiếu Turnstile token khi Turnstile bật. Turnstile missing token trả code TURNSTILE_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Đăng nhập không hợp lệ. Backend cố ý trả lỗi generic cho email không tồn tại, mật khẩu sai, tài khoản inactive, provider không phù hợp hoặc thiếu password hash. Không được dùng response này để trigger refresh.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "CSRF failure trả code CSRF_INVALID. Turnstile verification/config/provider failure trả code TURNSTILE_FAILED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Quá giới hạn request hoặc tạm khóa đăng nhập theo email/IP. Runtime hiện không đảm bảo machine-readable code hoặc countdown có cấu trúc.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Đăng nhập",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Thu hồi tất cả phiên đang hoạt động của user xác định từ cookie accessToken hoặc refreshToken, sau đó xóa cookie đăng nhập. Backend sử dụng accessToken hoặc refreshToken nếu cookie tồn tại để xác định user và revoke active sessions; chúng không phải security precondition bắt buộc.",
        "operationId": "authLogout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Đăng xuất thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Đăng xuất thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthMutationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "CSRF failure trả code CSRF_INVALID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Đăng xuất",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Cấp lại token mới từ refresh token trong cookie.",
        "operationId": "authRefresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Làm mới phiên đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Làm mới phiên đăng nhập thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthMutationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh session không hợp lệ. Có thể là missing refresh cookie, invalid/expired token, missing payload, missing/revoked/expired session, inactive user, REFRESH_TOKEN_ALREADY_ROTATED hoặc REFRESH_TOKEN_INVALID_OR_REUSED. REFRESH_TOKEN_ALREADY_ROTATED là concurrent rotation race loser; backend không revoke session và không clear cookies ở branch này. REFRESH_TOKEN_INVALID_OR_REUSED là refresh-token hash mismatch / invalid-or-reused; backend revoke matching session và clear auth cookies, không khẳng định đây là malicious reuse được xác nhận.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "alreadyRotated": {
                    "summary": "Concurrent rotation race loser",
                    "value": {
                      "statusCode": 401,
                      "message": "Refresh token đã được xoay vòng bởi một yêu cầu khác",
                      "error": "Unauthorized",
                      "code": "REFRESH_TOKEN_ALREADY_ROTATED",
                      "path": "/api/v1/auth/refresh",
                      "method": "POST",
                      "timestamp": "2026-07-02T00:00:00.000Z"
                    }
                  },
                  "invalidOrReused": {
                    "summary": "Refresh-token hash mismatch / invalid-or-reused",
                    "value": {
                      "statusCode": 401,
                      "message": "Refresh token không hợp lệ hoặc đã được dùng lại",
                      "error": "Unauthorized",
                      "code": "REFRESH_TOKEN_INVALID_OR_REUSED",
                      "path": "/api/v1/auth/refresh",
                      "method": "POST",
                      "timestamp": "2026-07-02T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "CSRF failure trả code CSRF_INVALID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Vượt quá giới hạn refresh trong cửa sổ throttle. Background refresh phải coi đây là transient/inconclusive, không tự động coi là session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshToken": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Làm mới phiên đăng nhập",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "authMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lấy thông tin người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy thông tin người dùng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthMeResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Access session không hợp lệ, hết hạn hoặc thiếu. Đây là eligible access-session 401 cho refresh flow nếu request chưa retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Lấy thông tin tài khoản hiện tại",
        "tags": [
          "auth"
        ],
        "description": "Lấy thông tin tài khoản hiện tại"
      }
    },
    "/auth/csrf-token": {
      "get": {
        "operationId": "authCsrfToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tạo CSRF token thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo CSRF token thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CsrfTokenResponseDto"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Không lấy được CSRF token do vượt throttle. Frontend không được gửi mutation phụ thuộc CSRF nếu bootstrap token thất bại.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Tạo CSRF token",
        "tags": [
          "auth"
        ],
        "description": "Tạo CSRF token",
        "security": []
      }
    },
...
    "/auth/google/callback": {
      "get": {
        "operationId": "authGoogleCallback",
        "parameters": [
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Chuyển hướng về frontend. Thành công: /auth/callback?success=true. Thất bại: /login?error=google_access_denied, /login?error=google_state_invalid hoặc /login?error=google_auth_failed."
          }
        },
        "summary": "Xử lý callback Google",
        "tags": [
          "auth"
        ],
        "description": "Xử lý callback Google",
        "security": []
      }
    }
  },
  "info": {
    "title": "Bookora API",
    "description": "Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.",
    "version": "1.0",
    "license": {
      "name": "Proprietary"
    }
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Cookie accessToken dùng để xác thực request cần user."
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Cookie refreshToken dùng cho refresh/logout phiên."
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csrfToken",
        "description": "Cookie csrfToken trong cơ chế CSRF double-submit. Phải đi cùng header X-CSRF-Token trên mutation cần CSRF."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Header CSRF double-submit. Giá trị phải trùng cookie csrfToken trên mutation cần CSRF."
      }
    },
    "schemas": {
...
      "UserType": {
        "type": "string",
        "enum": [
          "SYSTEM",
          "BRANCH",
          "CUSTOMER"
        ]
      },
      "PublicAuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T",
            "format": "ulid",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "minLength": 26,
            "maxLength": 26
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "customer@example.com"
          },
          "fullName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "type": {
            "example": "CUSTOMER",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserType"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "type"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "customer@example.com",
            "description": "Email dùng để đăng ký tài khoản."
          },
          "fullName": {
            "type": "string",
            "minLength": 2,
            "example": "Nguyen Van A",
            "description": "Họ tên hiển thị của user. Tối thiểu 2 ký tự."
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "pattern": "^(?=.*[A-Za-z])(?=.*\\d).+$",
            "example": "Password1",
            "description": "Mật khẩu đăng ký. Tối thiểu 8 ký tự và phải có ít nhất một chữ cái cùng một chữ số."
          },
          "turnstileToken": {
            "type": "string",
            "example": "0.M4AA-example-turnstile-token",
            "description": "Token lấy từ Cloudflare Turnstile widget trên frontend. Trường này là optional ở schema, nhưng bắt buộc khi Turnstile được bật trên server."
          }
        },
        "required": [
          "email",
          "fullName",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "customer@example.com",
            "description": "Email dùng để đăng nhập."
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "pattern": "^(?=.*[A-Za-z])(?=.*\\d).+$",
            "example": "Password1",
            "description": "Mật khẩu đăng nhập. Tối thiểu 8 ký tự và phải có ít nhất một chữ cái cùng một chữ số."
          },
          "turnstileToken": {
            "type": "string",
            "example": "0.M4AA-example-turnstile-token",
            "description": "Token lấy từ Cloudflare Turnstile widget trên frontend. Trường này là optional ở schema, nhưng bắt buộc khi Turnstile được bật trên server."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthMutationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
      "AuthMeRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "ulid",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "minLength": 26,
            "maxLength": 26,
            "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
          },
          "code": {
            "type": "string"
          },
          "level": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "SYSTEM",
              "BRANCH",
              "CUSTOMER"
            ]
          },
          "isSystem": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "code",
          "level",
          "type",
          "isSystem"
        ]
      },
      "AuthMeBranchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "ulid",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "minLength": 26,
            "maxLength": 26,
            "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "isPrimary"
        ]
      },
...
      "AuthMeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "ulid",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "minLength": 26,
            "maxLength": 26,
            "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "0901234567",
            "description": "Số điện thoại hồ sơ, null nếu chưa cập nhật."
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "example": "female",
            "description": "Giới tính dạng chuỗi hồ sơ, null nếu chưa cập nhật."
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "1995-06-15",
            "description": "Ngày sinh theo lịch, định dạng YYYY-MM-DD."
          },
          "type": {
            "type": "string",
            "enum": [
              "SYSTEM",
              "BRANCH",
              "CUSTOMER"
            ]
          },
          "roles": {
            "description": "Legacy request-effective roles. Empty for BRANCH users until a branch-scoped request resolves X-Branch-Id.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthMeRoleDto"
            }
          },
          "permissions": {
            "description": "Legacy request-effective permissions. Never contains a cross-branch union.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "globalRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthMeRoleDto"
            }
          },
          "globalPermissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "branchAssignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthMeBranchAssignmentDto"
            }
          },
          "maxRoleLevel": {
            "type": "number"
          },
          "isSuperAdmin": {
            "type": "boolean"
          },
          "branches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthMeBranchDto"
            }
          },
          "primaryBranchId": {
            "type": "string",
            "nullable": true,
            "format": "ulid",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "minLength": 26,
            "maxLength": 26,
            "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "phone",
          "gender",
          "birthday",
          "type",
          "roles",
          "permissions",
          "globalRoles",
          "globalPermissions",
          "branchAssignments",
          "maxRoleLevel",
          "isSuperAdmin",
          "branches",
          "primaryBranchId"
        ]
      },
      "CsrfTokenResponseDto": {
        "type": "object",
        "properties": {
          "csrfToken": {
            "type": "string",
            "example": "8f1f0d6c2ad34f6a9bb20b2ca3fc2a4f2c8873cfd5c5fb22ad91bc70ed9a6aa2"
          }
        },
        "required": [
          "csrfToken"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "required": [
          "statusCode",
          "message",
          "error",
          "path",
          "method",
          "timestamp"
        ],
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Dữ liệu không hợp lệ"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "code": {
            "type": "string",
            "example": "CSRF_INVALID"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": [
                "email must be an email"
              ]
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-28T12:00:00.000Z"
          }
        }
      }
    }
  }
}